Implement Ray Serve and BentoML support for wtpsplit sentence segmentation API

+++ wtpsplit/extract.py
@@ -72,11 +72,11 @@ class SaTTritonWrapper:
 
         # Prepare inputs for Triton
         inputs = []
-        inputs.append(grpcclient.InferInput("input_ids", input_ids.shape, "INT64"))
-        inputs.append(grpcclient.InferInput("attention_mask", attention_mask.shape, "INT64"))
+        inputs.append(grpcclient.InferInput("input_ids", input_ids.shape, "INT32"))
+        inputs.append(grpcclient.InferInput("attention_mask", attention_mask.shape, "INT32"))
 
-        inputs[0].set_data_from_numpy(input_ids.astype(np.int64))
-        inputs[1].set_data_from_numpy(attention_mask.astype(np.int64))
+        inputs[0].set_data_from_numpy(input_ids.astype(np.int32))
+        inputs[1].set_data_from_numpy(attention_mask.astype(np.int32))
 
         # Prepare outputs
         outputs = []
